refactor meta routes to separate module
